Bug fixes and 3rd party library updates

+++ src/main/java/io/mapsmessaging/storage/impl/file/FileHelper.java
@@ -1,18 +1,20 @@
 /*
- *   Copyright [2020 - 2022]   [Matthew Buckton]
  *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
+ *  Copyright [ 2020 - 2024 ] Matthew Buckton
+ *  Copyright [ 2024 - 2025 ] MapsMessaging B.V.
  *
- *        http://www.apache.org/licenses/LICENSE-2.0
+ *  Licensed under the Apache License, Version 2.0 with the Commons Clause
+ *  (the "License"); you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at:
  *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *      https://commonsclause.com/
  *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
  */
 
 package io.mapsmessaging.storage.impl.file;
@@ -23,6 +25,7 @@ import io.mapsmessaging.logging.LoggerFactory;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
+import java.util.concurrent.locks.LockSupport;
 
 import static io.mapsmessaging.storage.logging.StorageLogMessages.*;
 
@@ -69,13 +72,25 @@ public class FileHelper {
     }
     else{
       logger.log(FILE_HELPER_FILE_DOES_NOT_EXIST, file.toString());
-      throw new IOException("File does not exist");
+      throw new IOException("File does not exist "+file.toString());
     }
     return true;
   }
 
   private static boolean deleteExistingFile(File file) throws IOException {
-    Files.deleteIfExists(file.toPath());
+    int count =0;
+    while(count < 10){
+      if(file.exists()){
+        try {
+          Files.deleteIfExists(file.toPath());
+          count = 10;
+        } catch (IOException e) {
+          count++;
+          LockSupport.parkNanos(1000000);
+          if(count == 10)throw e;
+        }
+      }
+    }
     logger.log(FILE_HELPER_DELETED_FILE, file.toString());
     return true;
   }
